Create whodunit
The only changes from the given copy.c file was in lines 84~88. It takes pixels that have green & blue in them (green is not black, and blue is not black), and sets the green and blue to black. Red values remain at what they start at. Thus all-white values also turn into full-red.